Fix for Mac not accepting not override marked overriding methods (#1351)
* Fix for Mac not accepting not override marked overriding methods

* Update SelectionFilterAndAnalysis.cxx

+++ PWGCF/TwoParticleCorrelations/Core/SelectionFilterAndAnalysis.cxx
@@ -60,7 +60,7 @@ TString SelectionFilterAndAnalysis::printOptionalMasks() const
       str += ", ";
     }
     first = false;
-    str += TString::Format("0x%016lx", option);
+    str += TString::Format("0x%016lx", u_long(option));
   }
   str += ")";
   return str;
